Finished Test
Works better when focused on one room, still low fps however.

Had to fix multiple issues in order to get it to work. For some reason it seems the camera coordinate system seemed to be very twisted, the get layers function and the test game loop was both originally flipping the coordinates for some unknown reason which cancelled out and made it look fine but was horrible behind the scenes.

The coordinate system also used a mix of room and tile coordinates and converted to each other at multiple stages instead of using one consistent system. This mistake has been fixed and now it consistently uses tile coordinates as thats most relevant to the player.